ailab3/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=9.5.0
opencv-python>=4.5.0
matplotlib>=3.7.0
seaborn>=0.12.2